clean up Id tags, build system (add local build config)
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@20880 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ sixdeg/libsixdeg/encode_decode.cc
@@ -1,3 +1,4 @@
+/* $Id$ */
 /*
  * Six degrees of Wikipedia: Server (request handler).
  * This source code is released into the public domain.
@@ -48,7 +49,7 @@ request_encoder::send_to(int s)
 	}
 	buf.insert(buf.end(), eod.begin(), eod.end());
 
-	if (write(s, &buf[0], buf.size()) < buf.size())
+	if (write(s, &buf[0], buf.size()) < (ssize_t) buf.size())
 		return false;
 	return true;
 }
